Former-commit-id: 9e7e9c87aedc0634cb919e94a962f90d9a32bf10

+++ Tools/UWPRunner/Classes/AppxBundleHelper.cpp
@@ -56,9 +56,17 @@ AppxBundleHelper::AppxBundleHelper(const FilePath &fileName)
     }
 }
 
+void AppxBundleHelper::RemoveFiles()
+{
+    if (bundlePackageDir.Exists())
+    {
+        FileSystem::Instance()->DeleteDirectory(bundlePackageDir);
+    }
+}
+
 AppxBundleHelper::~AppxBundleHelper()
 {
-    FileSystem::Instance()->DeleteDirectory(bundlePackageDir);
+    RemoveFiles();
 }
 
 bool AppxBundleHelper::IsBundle(const FilePath &fileName)
